MotionBlockPyramid: widen nLambdaLevel to int64 to avoid overflow at plevel 2
In DoSearchMVs, nLambdaLevel (an int) is multiplied by nScale*nScale when
plevel == 2. On a deep pyramid nScale can be 8-16, so with a high
bit-depth-scaled lambda the product overflows 32-bit int before being
stored into the int64_t nLambda member, yielding a wrong (possibly
negative) lambda for that level. Make nLambdaLevel int64_t so the scaling
happens in 64-bit (the DoRecalculateMVs copy is widened too for
consistency, though it does no nScale scaling).

+++ src/MotionBlockPyramid.cpp
@@ -960,7 +960,7 @@ void MotionBlockLevel::DoSearchMVs(const FramePyramidLevel &pSrcFrame, const Fra
     searchType = st;    //( nLogScale == 0 ) ? st : EXHAUSTIVE;
     nSearchParam = stp; //*nPel; // v1.8.2 - redesigned in v1.8.5
 
-    int nLambdaLevel = lambda / ((1 << nLogPel) * (1 << nLogPel));
+    int64_t nLambdaLevel = lambda / ((1 << nLogPel) * (1 << nLogPel));
     if (plevel == 1)
         nLambdaLevel = nLambdaLevel * nScale; // scale lambda - Fizick
     else if (plevel == 2)
@@ -1164,7 +1164,7 @@ void MotionBlockLevel::DoRecalculateMVs(const FramePyramidLevel &pSrcFrame, cons
     searchType = st;
     nSearchParam = stp;
 
-    int nLambdaLevel = lambda / ((1 << nLogPel) * (1 << nLogPel));
+    int64_t nLambdaLevel = lambda / ((1 << nLogPel) * (1 << nLogPel));
 
 
     // Functions using float must not be used here
